Project restructuring. Transfer of functional parts to separate modules.
