+++ config/__init__.py
@@ -1,0 +1,2 @@
+from config.basic_config import BasicConfig
+from config.single_config import SingleConfig
