+++ field/__init__.py
@@ -1,0 +1,1 @@
+from field.fields import DefaultConfigField, ImmutableConfigField
