+++ file_worker/__init__.py
@@ -1,0 +1,1 @@
+from file_worker.types import such_a_type, use_list, config_to_string
